+++ packages/component/src/block/accountStatus/desktop/assets.tsx
@@ -107,7 +107,8 @@ export const Assets: FC<AssetsProps> = (props) => {
     return Math.min(
       10,
       aggregated.notional === 0
-        ? positionsInfo["margin_ratio"](10)
+        ? // @ts-ignore
+          positionsInfo["margin_ratio"](10)
         : marginRatio
     );
   }, [marginRatio, aggregated]);