+++ packages/core/__test__/account.test.ts
@@ -1,10 +1,10 @@
-import { Account } from "../src/account";
-import { memoryConfigStoreInstance } from "../src/configStore/configStore";
-import { SimpleWallet } from "../src/wallet";
+// import { Account } from "../src/account";
+// import { memoryConfigStoreInstance } from "../src/configStore/configStore";
+// import { SimpleWallet } from "../src/wallet";
 
-test("Account Class", () => {
-  const walletClient = new SimpleWallet(
-    "0xd659f3dece290110421a71b438b4f5a375f8be58"
-  );
-  // const account = new Account(memoryConfigStoreInstance,walletClient);
-});
+// test("Account Class", () => {
+//   const walletClient = new SimpleWallet(
+//     "0xd659f3dece290110421a71b438b4f5a375f8be58"
+//   );
+//   // const account = new Account(memoryConfigStoreInstance,walletClient);
+// });