+++ packages/net/__test__/createWebSocket.test.ts
@@ -1,11 +1,11 @@
-import WebSocketClient from "../src/ws";
+// import WebSocketClient from "../src/ws";
 
-describe("createWebSocket", () => {
-  let ws;
+// describe("createWebSocket", () => {
+//   let ws;
 
-  beforeAll(() => {
-    ws = new WebSocketClient({ url: "ws://localhost:8080" });
-  });
+//   beforeAll(() => {
+//     ws = new WebSocketClient({ url: "ws://localhost:8080" });
+//   });
 
-  test("should create a websocket", () => {});
-});
+//   test("should create a websocket", () => {});
+// });