+++ packages/perp/CHANGELOG.md
@@ -1,22 +1,64 @@
 # @orderly.network/futures
 
-## 2.0.7-referral.0
+## 3.0.0
+
+### Minor Changes
+
+- 52bebed3: fix bugs;
 
 ### Patch Changes
 
-- publish referral
+- Updated dependencies [52bebed3]
+  - @orderly.network/types@1.2.0
+  - @orderly.network/utils@1.2.0
+
+## 3.0.0-alpha.0
+
+### Minor Changes
+
+- fix bugs;
+
+### Patch Changes
+
 - Updated dependencies
-  - @orderly.network/types@0.2.7-referral.0
-  - @orderly.network/utils@0.1.7-referral.0
+  - @orderly.network/types@1.2.0-alpha.0
+  - @orderly.network/utils@1.2.0-alpha.0
+
+## 2.0.8
+
+### Patch Changes
+
+- fix bug: app crashes when symbolInfo is null;
+- Updated dependencies
+  - @orderly.network/types@0.2.8
+  - @orderly.network/utils@0.1.8
+
+## 2.0.7
+
+### Patch Changes
+
+- fix bug: window not defined;
+- Updated dependencies
+  - @orderly.network/types@0.2.7
+  - @orderly.network/utils@0.1.7
 
 ## 2.0.6
 
 ### Patch Changes
 
-- update version
-- Updated dependencies
+- 93a5ce6a: fix bugs;
+- Updated dependencies [93a5ce6a]
   - @orderly.network/types@0.2.6
   - @orderly.network/utils@0.1.6
+
+## 2.0.6-alpha.0
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/types@0.2.6-alpha.0
+  - @orderly.network/utils@0.1.6-alpha.0
 
 ## 2.0.5